Create new structs for trade data and quantity
